Warnings Be Gone

+++ src-core/modules/noaa/dsb_deframer.cpp
@@ -54,7 +54,7 @@ namespace noaa
         return state;
     }
 
-    std::vector<std::array<uint8_t, FRAME_SIZE>> DSBDeframer::work(uint8_t *input, size_t size)
+    std::vector<std::array<uint8_t, FRAME_SIZE>> DSBDeframer::work(uint8_t *input, int size)
     {
         // Output buffer
         std::vector<std::array<uint8_t, FRAME_SIZE>> frames;
